Return default values on broken configurations
Suppose that a configuration file contained a record
like this:

    CountSeconds = aaaa

Previously we returned 0, because we didn't bother to query
the default value when there is key found in `_options`.

This changes the behavior so that we fall back to the default
value (in this case, `3`) whenever the user value is broken,
which should return more safer/reasonable value to the caller.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>