Add default configuration files (#30)
* Add default configuration files

The default configuration files are Common.txt, TrustedDomains.txt, UnsafeDomains.txt and UnsafeFiles.txt in `{app}\DefaultConfig`.
Those files are installed when they are located in the "DefaultConfig" folder that exits in the same folder to installer like below.

```
│  FlexConfirmMailSetup-22.4.0.exe
│
└─DefaultConfig
        Common.txt
        TrustedDomains.txt
        UnsafeDomains.txt
```

The configurations in the policy configuration and the default configuration files for trusted domains, unsafe domains and unsafe files
are fixed, they can't modified by user configuration files.
The priority of common configurations is as following.

User file configurations > Default file configurations > Policy configurations