Fix a bug that deferred delivery is blocked if a main in outbox is selected.
When deferred delivery is enabled, a mail is moved to the outbox and delivered after spending the delay time.
If we access any of properties of the mail in the outbox, the mail is regarded as "edited" and not delivered
even after spending the delay time.
So if the folder is the outbox, return before accessing properties of the mail.
Note that if a mail in the outbox is opened with a new inspector, it is regarded as "edited", which is Outlook's
specification, so we don't add this check for ThisAddIn_NewInspector.