Avoid custom UI error
The "TabMail" tab exists only in the ribbon whose id is "Microsoft.Outlook.Explorer". So if we open a ribbon
whose id is not "Microsoft.Outlook.Explorer", we may get a custom UI error like "TabMail" doesn't exist.
The error is displayed when Outlook's setting "Options" -> "Advanced" -> "Show add-in user interface errors"
is enabled.

This patch avoid that error by creating custom UI only when ribbonID is "Microsoft.Outlook.Explorer".